+++ packages/nimble-components/src/menu/styles.ts
@@ -12,7 +12,8 @@ import {
     groupHeaderFontColor,
     smallPadding,
     mediumPadding,
-    elevation2BoxShadow
+    elevation2BoxShadow,
+    menuMinWidth
 } from '../theme-provider/design-tokens';
 import { Theme } from '../theme-provider/types';
 import { hexToRgbaCssColor } from '../utilities/style/colors';
@@ -25,7 +26,7 @@ export const styles = css`
         background: ${applicationBackgroundColor};
         border: ${borderWidth} solid ${popupBorderColor};
         margin: 0;
-        min-width: 176px;
+        min-width: ${menuMinWidth};
         width: max-content;
         box-shadow: ${elevation2BoxShadow};
     }
