+++ packages/nimble-components/src/theme-provider/design-token-names.ts
@@ -53,6 +53,7 @@ export const tokenNames: { readonly [key in TokenName]: string } = {
     dialogLargeWidth: 'dialog-large-width',
     dialogLargeHeight: 'dialog-large-height',
     dialogLargeMaxHeight: 'dialog-large-max-height',
+    menuMinWidth: 'menu-min-width',
     bannerGapSize: 'banner-gap-size',
     spinnerSmallHeight: 'spinner-small-height',
     spinnerMediumHeight: 'spinner-medium-height',
